+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   initialized
 - Rename `VSC` &rarr; `Stats`, `VSCBuilder` &rarr; `StatsBuilder`, and `VSCInternal` into `StatsImpl`
 - Consolidate `Probe` and `CowProbe<'a>` into one `Probe<T>` struct with a generic `String` or `Cow<str>`.
+- Rename `new` to `from_ptr` for `Buffer`, `Workspace`, `HttpHeaders`, `FetchProcCtx`, `DeliveryProcCtx`, and make them private. Only `Ctx::from_ptr` is public because it gets created by a macro-generated function.
 
 # 0.1.0 (2024-11-12)